feat: 분리된 프로필탭에 css 적용
