+++ MUDIUM_Frontend/src/assets/styles/typography.css
@@ -11,7 +11,7 @@
 
 h1 {
   font-size: 3rem;
-  color: var(--primary-color);
+  color: var(--text-color);
 }
 
 h2 {
